Pequeño movimiento de función en calendario dialog
