+++ src/sections/auth/jwt/jwt-register-view.jsx
@@ -1,6 +1,5 @@
 import * as Yup from 'yup';
 import { useState } from 'react';
-import { Base64 } from 'js-base64';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { useNavigate, useLocation } from 'react-router';
@@ -16,17 +15,14 @@ import useMediaQuery from '@mui/material/useMediaQuery';
 
 import { RouterLink } from 'src/routes/components';
 import { useRouter, useSearchParams } from 'src/routes/hooks';
-import { useSnackbar } from 'src/components/snackbar';
 
 import { useBoolean } from 'src/hooks/use-boolean';
 
-import instance from 'src/utils/axiosCH';
-
+import { getColaborador } from 'src/api/user';
 import { useAuthContext } from 'src/auth/hooks';
 import { PATH_AFTER_LOGIN, PATH_AFTER_REGISTRO } from 'src/config-global';
 
-import { getColaborador } from 'src/api/user';
-
+import { useSnackbar } from 'src/components/snackbar';
 import FormProvider, { RHFTextField } from 'src/components/hook-form';
 
 import ModalPoliticas from './modal-politicas';
@@ -35,7 +31,7 @@ import ModalPoliticas from './modal-politicas';
 export default function JwtRegisterView() {
   const [numEmpleado, setnumEmpleado] = useState('');
 
-  const { enqueueSnackbar } = useSnackbar()
+  const { enqueueSnackbar } = useSnackbar();
 
   const { register } = useAuthContext();
 
@@ -89,50 +85,21 @@ export default function JwtRegisterView() {
   });
 
   const validarNumEmpleado = async () => {
-      //  console.log('Ingresar número de empleado válido');
-      // Conectar axios con CH
-      // const datos = Base64.encode(JSON.stringify({ num_empleado: numEmpleado })); // VALLE01392
+    const response = await getColaborador(numEmpleado);
+    if (!response.result) {
+      enqueueSnackbar('Número de empleado no encontrado', { variant: 'error' });
+      return false;
+    }
+    navigate(PATH_AFTER_REGISTRO, { state: response });
+    location(PATH_AFTER_REGISTRO, { state: response });
+    router.push(returnTo || PATH_AFTER_REGISTRO);
 
-      // const config = {
-      //   headers: {
-      //     Authorization: '41EgSP8+YSqsyT1ZRuxTK3CYR11LOcyqopI2TTNJd3EL+aU3MUdJNsKGx8xOK+HH',
-      //     Accept: '*/*',
-      //     Origin: 'https://prueba.gphsis.com/auth/jwt/register',
-      //   },
-      // };
-
-      const response = await getColaborador(numEmpleado);
-      if (!response.result) {
-        enqueueSnackbar('Número de empleado no encontrado', { variant: 'error' });
-        return false;
-      }
-      // alert(JSON.stringify(response.data[0]));
-      navigate(PATH_AFTER_REGISTRO, { state: response });
-      location(PATH_AFTER_REGISTRO, { state: response });
-      router.push(returnTo || PATH_AFTER_REGISTRO);
-      // instance
-      //   .post('https://rh.gphsis.com/index.php/WS/data_colaborador_consultas', datos, config)
-      //   .then((response) => {
-      //     let datosResponse = Base64.decode(JSON.stringify(response.data));
-      //     datosResponse = JSON.parse(datosResponse);
-      //     // if(datosResponse.resultado === 0){ // se cambio ya que este no daba el resultado
-      //     if (datosResponse.data.length === 0) {
-      //       setErrorMsg('Número de empleado no encontrado');
-      //     } else {
-      //       navigate(PATH_AFTER_REGISTRO, { state: datosResponse });
-      //       location(PATH_AFTER_REGISTRO, { state: datosResponse });
-      //       router.push(returnTo || PATH_AFTER_REGISTRO);
-      //     }
-      //   })
-      //   .catch((error) => {
-      //     console.error(error);
-      //   });
-
-      return true;
+    return true;
   };
 
   const renderHead = (
-    <Stack spacing={1} sx={{ mb: 0 }}>´
+    <Stack spacing={1} sx={{ mb: 0 }}>
+      ´
       <Box
         component="img"
         alt="auth"
@@ -142,23 +109,22 @@ export default function JwtRegisterView() {
           position: 'absolute',
           width: { xs: '25%', md: '16%' },
           left: '64%',
-          top: { xs: '10%', md: '-9%' }
+          top: { xs: '10%', md: '-9%' },
         }}
       />
       <Box
-          component="img"
-          alt="auth"
-          src={`${import.meta.env.BASE_URL}assets/img/beneficiosBrand.svg`}
-          sx={{
-            maxWidth: {
-              xs: 480,
-              lg: 560,
-              xl: 720,
-            }
-          }}
-        />
+        component="img"
+        alt="auth"
+        src={`${import.meta.env.BASE_URL}assets/img/beneficiosBrand.svg`}
+        sx={{
+          maxWidth: {
+            xs: 480,
+            lg: 560,
+            xl: 720,
+          },
+        }}
+      />
       <Typography variant="h4">Registro de usuarios</Typography>
-
       <Stack direction="row" spacing={0.5}>
         <Typography variant="body2"> ¿Ya tienes una cuenta? </Typography>
 
@@ -184,7 +150,7 @@ export default function JwtRegisterView() {
 
       {'Al registrarme, acepto '}
       {' las '}
-      <Button variant="outlined" color="primary" sx={{height:"20px"}} onClick={quickEdit.onTrue}>
+      <Button variant="outlined" color="primary" sx={{ height: '20px' }} onClick={quickEdit.onTrue}>
         Políticas de privacidad
       </Button>
     </Typography>
@@ -219,32 +185,32 @@ export default function JwtRegisterView() {
   const logoMd = (
     <Stack>
       {isMobile && (
-          <Box
-            component="img"
-            alt="auth"
-            src={`${import.meta.env.BASE_URL}assets/img/logoMaderas.svg`}
-            sx={{
-              maxWidth: { xs: 480, lg: 560, xl: 720 },
-              position: 'absolute',
-              width: { xs: '50%', md: '21%' },
-              left: '25%',
-              top: { xs: '83%'}
-            }}
-          />
-        )}
-        {!isMobile && (
-          <Box
-            component="img"
-            alt="auth"
-            src={`${import.meta.env.BASE_URL}assets/img/logoMaderas.svg`}
-            sx={{
-              position: 'absolute',
-              left: '37%',
-              width: { xs: '55%', md: '26%' },
-              top: { md: '103%' }
-            }}
-          />
-        )}
+        <Box
+          component="img"
+          alt="auth"
+          src={`${import.meta.env.BASE_URL}assets/img/logoMaderas.svg`}
+          sx={{
+            maxWidth: { xs: 480, lg: 560, xl: 720 },
+            position: 'absolute',
+            width: { xs: '50%', md: '21%' },
+            left: '25%',
+            top: { xs: '83%' },
+          }}
+        />
+      )}
+      {!isMobile && (
+        <Box
+          component="img"
+          alt="auth"
+          src={`${import.meta.env.BASE_URL}assets/img/logoMaderas.svg`}
+          sx={{
+            position: 'absolute',
+            left: '37%',
+            width: { xs: '55%', md: '26%' },
+            top: { md: '103%' },
+          }}
+        />
+      )}
     </Stack>
   );
